Limit maxGxpDebugDumps to 3.
DSP multicore workload can run atmost on 3 cores thus in case
of crash we will need atmost 3 core dumps. Thus limiting
maxGxpDebugDumps to 3 to make sure that we will be capturing
all the required core dumps for the recent crash and at the
same time not inflating the bugreport size.

Bug: 275756225,274956438
Test: Verified the changes on the device using ab/P53879006.
Change-Id: Iede8ec4204d58fdedbe4141bf57fc434b411bf0c
Signed-off-by: nischaljain <nischaljain@google.com>