Add SEPolicy config for PCS to act as a media app
These are the minimum set of services that PCS needs to have access for
it to be able act as a media app and use Exoplayer for playing recorded
video files.

However, there'll be a follow up change to broaden the permissions to be
future proof and have greater flexibility as a media app, which will let
PCS to be updated via Play Store without the worry of a missing SEPolicy
config that is common among media apps.

Bug: 287069860
Test: m && flashall
Change-Id: I956219faacbc0c1b649cb638cede964480766718