Move definition for vendor_camera_binder_service into product
vendor_camera_binder_service is defined in vendor image, but this
property is required and used from the system image. This causes
Cuttlefish Hybrid Device to fail from sepolicy error. This change is to
move system-required property from vendor to product so it can be used
when vendor image is changed into generic one.

Bug: 309469924
Test: Build and boot succeeded with cheetah
Change-Id: Iea3e5be110498f759e268df8b7e5126b65b06a67