dump_modemlog: always move modem logs
Dynamic log mask events can occur without leaving the logging status
property as enabled, which means when dumpstate should always try to
stop modem logging so that the new logs can be copied over.

Having the copying of logs and stopping of modem logging combined in
one command is no longer an ideal design, so b/289435256 was created
to find a better solution to this.

Test: build, flash, trigger log mask event, check logs in bugreport.
Bug: 302435001
Change-Id: I56358d3f08ac1f2a6099ede14c5e17b5ebffabbd